% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revealLocation.R
\name{revealLocation}
\alias{revealLocation}
\title{\code{revealLocation} Revealing locations encoded as a string calculation.}
\usage{
revealLocation(hiddenLocation)
}
\arguments{
\item{hiddenLocation}{A character vector of calculation strings.}
}
\value{
Returns the value of the each calculation.
}
\description{
Reveals the location by parsing and evaluating the string.
}
\examples{
trueLoc <- hideLocation(3,100)
trueLoc
revealLocation(ans)

n <- 200
trueLoc <- sample(1:n, 3)
trueLoc
ans <- hideLocation(trueLoc , n)
ans
revealLocation(ans)

}
\seealso{
\code{\link{hideLocation}}
}
